Removed MPI from tests, reorganized tests
